% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FeatureMatrix_SQLite.R
\name{add_FeatureMatrix_SQLite}
\alias{add_FeatureMatrix_SQLite}
\title{Adding feature matrix table to the sqlite database}
\usage{
add_FeatureMatrix_SQLite(dbfile, xcms_exp)
}
\arguments{
\item{dbfile}{\code{character(1)} the path to the SQL database.}

\item{xcms_exp}{\code{character(1)} the path to the XCMSExperiment object.}
}
\value{
Updated SQL database with feature matrix table
}
\description{
Adding feature matrix table to the sqlite database
}
\details{
Adding feature matrix table which contains the feature intensities in each
sample to the sqlite database
}
\author{
Ahlam Mentag
}
