% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{add_ms3_ms4_candidates}
\alias{add_ms3_ms4_candidates}
\title{Generate MS3/MS4 candidates per feature}
\usage{
add_ms3_ms4_candidates(sd_best, sd_filtered, ms4_map)
}
\arguments{
\item{sd_best}{data.frame best MS trees}

\item{sd_filtered}{data.frame filtered spectra}

\item{ms4_map}{data.frame MS4 mapping table}
}
\value{
data.frame candidate spectra
}
\description{
Expands MS3 candidates and links MS4 spectra when available.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
