% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{best_tree_pipeline}
\alias{best_tree_pipeline}
\title{Run full MS tree selection pipeline}
\usage{
best_tree_pipeline(ftms_msn_tree)
}
\arguments{
\item{ftms_msn_tree}{Spectra object}
}
\value{
list with filtered data and selected FTMS object
}
\description{
Executes filtering, deduplication and best tree selection.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
