% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.GNPS_SQL.R
\name{conv.GNPS_SQL}
\alias{conv.GNPS_SQL}
\title{Convert and rank GNPS-like MS/MS relationships and store results
in SQLite}
\usage{
conv.GNPS_SQL(
  sql_path,
  expid,
  peakwidth = 0.2,
  mzerr = 0.015,
  min = 5,
  adduct = 46.0055,
  thr1 = 3,
  thr2 = 0.1,
  thr3 = 0.4,
  spectrum_type = "assembled",
  IntThres = 5
)
}
\arguments{
\item{sql_path}{\code{Character(1)} string giving the path to the SQLite database.}

\item{expid}{Integer. Experiment ID used to select compounds from the database.}

\item{peakwidth}{\code{Numeric(1)} half of the retention time window that is
defined for the CSPP ‘substrate’ feature allowing
a minimum retention time difference with the CSPP ‘product’ feature,
by default set on \code{0.2}.}

\item{mzerr}{\code{Numeric(1)} mass tolerance (in Da) used to match precursor and
product ions. Default it is configured for ftms data
\code{0.01}, for qtof data the user could set it to \code{0.015}.}

\item{min}{minimum absolute intensity of product ion to be withheld for CID
spectral matching, by default \code{5} for QTOF, but the user can set it to
\code{100} for FTMS spectra.}

\item{adduct}{mass of buffer, by default \code{46.0055 Da} representing
formic acid.}

\item{thr1}{average of the number of common ions and the number of common
losses, set on \code{3} by default.}

\item{thr2}{ratio of thr1 to the number of product ions in the CID spectrum
of the feature in the feature pair having the lowest number of product ions,
by default set on \code{0.1}.}

\item{thr3}{average of the dot products obtained for the common product ions
and the common neutral losses, by default set on \code{0.4}.}

\item{IntThres}{Numeric intensity threshold applied to MS/MS fragment ions
before similarity calculations. Default it is configured for ftms data
\code{100}, for qtof data the user could set it to \code{5}.}
}
\value{
Invisibly returns a data frame corresponding to the updated GNPS SQL table
for the processed experiment.
}
\description{
This function iterates the GNPS-like feature pair generation procedure over
all compound_id belonging to an experiment. the features (cfr. compound_id)
are first sorted with increasing retention times.
For each feature, CID spectral similarities are computed with all features
that elute later. Only those feature pairs for which the CID spectral
similarity passes certain thresholds (thr1, thr2 and thr3) are retained.
Subsequently, for a particular feature, all generated feature pairs are
ranked, via the rank.GNPS() function.
}
\author{
Ahlam Mentag
}
