% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpectraSQLite.R
\name{createSpectraSQLite}
\alias{createSpectraSQLite}
\title{Create new SQL database from a given spectra object}
\usage{
createSpectraSQLite(
  sps,
  dbfile,
  date,
  user,
  machine,
  mode,
  tissue,
  mstype = NULL,
  column = NULL,
  buffera = NULL,
  bufferb = NULL,
  gradient_time = NULL,
  source = NULL,
  species = NULL,
  ce = NULL,
  meta = NULL
)
}
\arguments{
\item{sps}{\code{character(1)} the path to the spectra object to convert.}

\item{dbfile}{\code{character(1)} the path to the output sql database.}

\item{date, }{user, machine, mode, tissue, mstype, column, buffera,
bufferb, gradient_time, source, species, ce, meta
Experimental metadata provided by the user. These fields describe
the acquisition conditions, sample origin, and experimental setup
associated with the spectra.}
}
\value{
SQL database containing the spectral information of the given object
}
\description{
Create new SQL database from a given spectra object
}
\details{
From a spectra object createSpectraSQLite() function creates an sqlite
database with the spectra data of the object given in the input of the
function, the resulting sql structure is compatible with 'MsBackendCompDb'
from 'RforMassSpecrtrometry' ecosystem, this allows to convert the sql
data easily to a spectra object.
}
\author{
Ahlam Mentag
}
