% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{deduplicate_ms3}
\alias{deduplicate_ms3}
\title{Deduplicate MS3 spectra}
\usage{
deduplicate_ms3(sd)
}
\arguments{
\item{sd}{data.frame containing MS spectra metadata}
}
\value{
data.frame deduplicated spectra
}
\description{
Removes redundant MS3 spectra by keeping the most intense one per group.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
