% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_similarity_helperfunctions.R
\name{dynlib_symmetric_dotproduct}
\alias{dynlib_symmetric_dotproduct}
\title{Calculate spectral similarity using the DynLib symmetric dot product}
\usage{
dynlib_symmetric_dotproduct(x, y, n = 3, m = 0.6, ...)
}
\arguments{
\item{x}{\code{numeric} 2-column \code{matrix} with the m/z and intensity values of
\strong{matched} peaks.}

\item{y}{\code{numeric} 2-column \code{matrix} with the m/z and intensity values of
\strong{matched} peaks.}

\item{n}{\code{numeric}; exponent applied to intensities in the weighted dot-product calculation.}

\item{m}{\code{numeric}; exponent applied to m/z values in the weighted dot-product calculation.}
}
\value{
a \code{numeric(1)} with the similarity between the matched peaks.
}
\description{
Function to calculate the symmetric dot product between two spectra. As
input \strong{matched} peak matrices are expected, such as those returned by
\code{dynlib_map()}. These input matrices are expected to have the same number
of rows, with the same row in both matrices \code{x} and \code{y} representing a
matching peak. The matrices can also contain neutral loss peaks. Importantly
each matrix has to also contain the m/z-weighted intensity sum as an
\emph{attribute} to the matrix (i.e. \verb{attributes(x)$wintensity_sum)}).
}
\author{
Ahlam Mentag
}
