% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_similarity_helperfunctions.R
\name{dynlibmatch2_map}
\alias{dynlibmatch2_map}
\title{Match fragment and neutral-loss peaks between spectra}
\usage{
dynlibmatch2_map(
  x,
  y,
  xPrecursorMz,
  yPrecursorMz,
  n = 3,
  m = 0.6,
  fragments_method = c("unit.resolution", "high.resolution"),
  tolerance = 0,
  ppm = 0,
  ...
)
}
\value{
a list with numeric matrices containing only the matched peaks between
the two input spectra. The m/z weighted intensity sum of each (cleaned) peak
matrix is returned as an attribute \code{"wintensity_sum"} to each table.
}
\description{
Matches/maps peaks between the two provided peak matrices.
}
\author{
Ahlam Mentag
}
