% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{enrich_ms3_ms4}
\alias{enrich_ms3_ms4}
\title{Enrich FTMS trees with MS3/MS4 spectra}
\usage{
enrich_ms3_ms4(ftms_best, sd_filtered)
}
\arguments{
\item{ftms_best}{Spectra object}

\item{sd_filtered}{data.frame filtered spectra}
}
\value{
Spectra object enriched
}
\description{
Rebuilds Spectra object with MS3/MS4 enrichment.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
