% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{filter_intense_prec}
\alias{filter_intense_prec}
\title{Filter MS trees by MS2 precursor intensity}
\usage{
filter_intense_prec(sd, quantile_cut = 0.25)
}
\arguments{
\item{sd}{data.frame containing MS spectra metadata}

\item{quantile_cut}{numeric(1) quantile cutoff (default 0.25)}
}
\value{
data.frame filtered spectra
}
\description{
Keeps MS trees whose MS2 precursor intensities are above a quantile threshold.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
