% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound.R
\name{get_compound}
\alias{get_compound}
\title{extracting the ms_compound table of a given compound_id}
\usage{
get_compound(qt_path, compound_id)
}
\value{
the ms_compound table of a given compound_id
}
\description{
extracting the ms_compound table of a given compound_id
}
\author{
Ahlam Mentag
}
