% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectrum_id.R
\name{get_spectrum_id}
\alias{get_spectrum_id}
\title{extracting the msms_spectrum table of a given compound_id}
\usage{
get_spectrum_id(qt_path, compound_id)
}
\value{
the msms_spectrum table of a given compound_id
}
\description{
extracting the msms_spectrum table of a given compound_id
}
\author{
Ahlam Mentag
}
