% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNegPos_SQL.R
\name{matchNegPos_SQL}
\alias{matchNegPos_SQL}
\title{Match  MS2 spectra between the same instruments type
using SQL databases.}
\usage{
matchNegPos_SQL(
  LCal,
  FTn_con,
  FTp_con,
  polarity_ftn = 0,
  polarity_ftp = 1,
  minIon = 0.2,
  tol = 0.01
)
}
\arguments{
\item{LCal}{\code{data.frame} returned by \code{Aligning_General_SQL()} function.}

\item{FTn_con}{\code{DBIConnection}
A \verb{DBI connection} object to the reference SQLite database.}

\item{FTp_con}{\code{DBIConnection}
A \verb{DBI connection} object to the SQLite database to align with.}

\item{polarity_ftn}{\code{Integer (0/1)} Polarity of negative FTMS or QTOF experiment.}

\item{polarity_ftp}{\code{Integer (0/1)} Polarity of positive FTMS or QTOF experiment.}

\item{minPeaks}{\code{Numeric(1)} Minimum matching ratio (default 0.2).}
}
\value{
Filtered \code{LCal} with MS2-supported neg–pos pairs.
}
\description{
\code{matchNegPos_SQL()} Removes false positive negFTm/z,posFTm/z peak pairs by
}
