% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMSspecplot_SQL.R
\name{MSMSspecplot_SQL}
\alias{MSMSspecplot_SQL}
\title{plot MS2 spectra}
\usage{
MSMSspecplot_SQL(
  sql_path,
  spectrum_id,
  prdion,
  neutloss,
  err = NULL,
  minum = NULL,
  nr_col = NULL,
  nr_col2 = NULL
)
}
\description{
plot MS2 spectra
}
\author{
Ahlam Mentag
}
