% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultMatchCID_Spectra.R
\name{MultMatchCID_Spectra}
\alias{MultMatchCID_Spectra}
\title{Match a query spectrum against spectra in a DynLib database}
\usage{
MultMatchCID_Spectra(
  spectra_db,
  compound_id,
  polarity_query,
  polarity_target,
  spectrum_type_query = NULL,
  spectrum_type_target = NULL,
  ms_level_query = NULL,
  ms_level_target = NULL,
  minIons = 3,
  fragments_resolution = c("unit.resolution", "high.resolution"),
  requirePrecursor = TRUE,
  threshold = 0.8,
  ppm = 2,
  tolerance = 0.5
)
}
\arguments{
\item{spectra_db}{\code{Spectra} object, represents the whole database.}

\item{compound_id}{compound_id to compare with the compounds from spectra_db.}

\item{polarity_query}{'numeric' the query polarity, 0  if negative,
and 1 for positive polarity.}

\item{polarity_target}{'numeric' the target polarity, 0  if negative,
and 1 for positive polarity.}

\item{fragments_resolution}{'character(1)' the resolution type of the fragments
masses, it is used to define the rounding type of the fragments,
it could be either "unit.resolution" or "high.resolution".}

\item{requirePrecursor}{'logical(1)' by default true, it allows to pre-filter
the target spectra prior to the actual similarity calculation for
each individual query spectrum.}

\item{threshold}{'numeric(1)' the score threshold to select a best match, by
default it is 0.8.}

\item{ppm}{'numeric(1)' the acceptable difference between m/z values of the
compared peaks:
\itemize{
\item For high resolution : ppm is applied for matching the product ions
and match precursorMz if requirePrecursor is TRUE.
\item For unit resolution : ppm is used for matching precursorMz.
}}

\item{tolerance}{'numeric(1)' the acceptable difference between m/z values
of the compared peaks.}
}
\value{
a matrix with the target and the query matches spectraData, the score, and
the number of matched peaks.
}
\description{
Calculates spectral similarity between a query spectrum, selected by
\code{compound_id}, and candidate spectra in a DynLib \code{Spectra} database.
The function supports both high-resolution and unit-resolution MS/MS data
and applies resolution-specific peak matching and similarity scoring
methods.
}
\author{
Ahlam Mentag
}
