% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodenameMass_SQL.R
\name{NodenameMass_SQL}
\alias{NodenameMass_SQL}
\title{extract compounds based on their nodename, compound name, and
precursorMz in a given experiment.}
\usage{
NodenameMass_SQL(
  sql_path,
  expid = NULL,
  nodename = NULL,
  name = NULL,
  mass_measured = NULL,
  rt = NULL,
  mass_range = 0.02,
  rt_range = 1
)
}
\arguments{
\item{sql_path}{'character(1)' path to the sqlite database.}

\item{expid}{'numeric(1)' experiment number.}

\item{nodename}{'character(1)' corresponds to the feature_id or nodename in
the format of "MxTy".}

\item{name}{'character(1)'  name or a part of the name of the compounds.}

\item{mass_measured}{'numeric(1)' the measured mass of the compounds.}

\item{mass_range}{the window error of the compound mass.}

\item{rt_range}{the window error of the compound retention time.}
}
\value{
It returns the rows of the compound table that match the query.
}
\description{
The function NodenameMass_SQL() look for compounds in a given
sqlite database. It lets the user query compounds in three possible ways:
\enumerate{
\item by nodename.
\item by compound name (partial match)
\item by m/z value within a tolerance
}
}
\author{
Ahlam Mentag
}
