% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall.net_SQL.R
\name{Overall.net_SQL}
\alias{Overall.net_SQL}
\title{Local networks for a selected COMPID}
\usage{
Overall.net_SQL(
  sql_path,
  exp.id,
  dbkey,
  min,
  nr_of_seq = 2,
  thr1 = 0,
  thr2 = 0,
  thr3 = 0
)
}
\arguments{
\item{sql_path}{'character(1)' path to the sqlite database.}

\item{exp.id}{'number(1)' experiment to use for network generation.}

\item{dbkey}{'number(1)' compound_id, serving as the starting point of the
molecular networking.Only nodes and edges connected to this COMPID within
the specified number of sequential connections (\code{nr_of_seq}) are included
in the displayed CSPP and GNPS local networks.}

\item{min}{numeric(1) Minimum intensity for the spectra.
For instruments with high sensitivity, a value like 5 can be used, whereas
for instruments with lower sensitivity, a value like 100 may be appropriate.}

\item{nr_of_seq}{'number(1)' size of local network, i.e., number of
subsequent edges starting from the node representing the selected COMPID.}

\item{thr1}{'number(1)' minimum number of product ions (varying between 0
and 1) of one CID spectrum that can be traced in another CID spectrum,
set on 1 by default.}

\item{thr2}{'number(1)' dot product threshold for the common ions
between two CID spectra, set on 0.9 by default.}

\item{thr3}{'number(1)' average of the dot products obtained for the common
product ions and the common neutral losses, by default set on 0.4.}
}
\value{
returns a list of the cspp and gnps-like information and plots
the final network of the given compound_id in a given experiment.
}
\description{
the Overall.net_SQL() function plots the CSPP or GNPS-like
local networks, it iterates through the net.addit() and Local.net_SQL()
functions. The net.addit() function is responsible for generating lists
of network nodes and edges by the net.nodes() and either the allCSPP() or
allGNPS() functions, respectively. This generates the nodes and edges for
the experiment-based global network. Only edges are retained that pass
certain CID spectral similarity thresholds. The Local.net_SQL() function will
select those nodes and edges belonging to the local network of the selected
COMPID and display the network graph.
}
\author{
Ahlam Mentag
}
