% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_perl.R
\name{round_perl}
\alias{round_perl}
\title{Round a number using Perl-style rounding}
\usage{
round_perl(number)
}
\arguments{
\item{number}{A numeric value to be rounded.}
}
\value{
An integer rounded from \code{number}.
}
\description{
This function rounds a numeric value to the nearest integer using
Perl-style rounding (i.e., adding 0.5 and taking the floor).
}
\examples{
round_perl(2.3)
round_perl(2.7)
}
