% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{select_best_ms3ms4}
\alias{select_best_ms3ms4}
\title{Select best MS3 and optional MS4 spectra}
\usage{
select_best_ms3ms4(ms3_df, sd_filtered)
}
\arguments{
\item{ms3_df}{data.frame MS3 candidates}

\item{sd_filtered}{data.frame full filtered spectra}
}
\value{
data.frame enriched spectra
}
\description{
Selects best MS3 spectrum and attaches MS4 if available.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
