% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{select_best_tree}
\alias{select_best_tree}
\title{Select best MS tree per feature}
\usage{
select_best_tree(sd, longest_trees)
}
\arguments{
\item{sd}{data.frame containing MS spectra metadata}

\item{longest_trees}{data.frame output of select_longest_trees()}
}
\value{
data.frame selected trees
}
\description{
Selects best MS tree based on peak count and precursor intensity.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
