% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_ftms_tree_pipeline.R
\name{select_longest_trees}
\alias{select_longest_trees}
\title{Select longest MS tree per feature}
\usage{
select_longest_trees(sd)
}
\arguments{
\item{sd}{data.frame containing MS spectra metadata}
}
\value{
data.frame selected trees
}
\description{
Selects the longest fragmentation tree for each feature.
}
\author{
Ahlam Mentag, Rebecca Dauwe
}
