% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_RDynLib_match.R
\name{similarity_RDynLib_match}
\alias{similarity_RDynLib_match}
\title{Search a DynLib spectral database for matches to query spectra}
\usage{
similarity_RDynLib_match(
  st_sps,
  dy_sps,
  polarity_query,
  polarity_target,
  fragments_resolution = c("unit.resolution", "high.resolution"),
  requirePrecursor = TRUE,
  threshold = 0.8,
  ppm = 2,
  tolerance = 0.5
)
}
\arguments{
\item{st_sps}{\code{Spectra} object, represents the query data.}

\item{dy_sps}{\code{Spectra} object, represents the target data.}

\item{polarity_query}{'numeric' the query polarity, 0  if negative,
and 1 for positive polarity.}

\item{polarity_target}{'numeric' the target polarity, 0  if negative,
and 1 for positive polarity.}

\item{fragments_resolution}{'character(1)' the resolution type of the fragments
masses, it is used to define the rounding type of the fragments,
it could be either "unit.resolution" or "high.resolution".}

\item{requirePrecursor}{'logical(1)' by default true, it allows to pre-filter
the target spectra prior to the actual similarity calculation for
each individual query spectrum.}

\item{threshold}{'numeric(1)' the score threshold to select a best match, by
default it is 0.8.}

\item{ppm}{'numeric(1)' the acceptable difference between m/z values of the
compared peaks:
\itemize{
\item For high resolution : ppm is applied for matching the product ions
and match precursorMz if requirePrecursor is TRUE.
\item For unit resolution : ppm is used for matching precursorMz.
}}

\item{tolerance}{'numeric(1)' the acceptable difference between m/z values
of the compared peaks.}
}
\value{
a matrix with the target and the query matches spectraData, the score, and
the number of matched peaks.
}
\description{
Compares spectra from a query \code{Spectra} object, typically representing a
newly acquired experiment, against a target \code{Spectra} object derived from
a DynLib SQL database to identify spectral matches and transfer compound
annotations. The function supports both high-resolution and unit-resolution
mass spectrometry data and applies resolution-specific spectral matching
and similarity scoring methods.
}
\author{
Ahlam Mentag
}
