% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_to_mgf.R
\name{sql_to_mgf}
\alias{sql_to_mgf}
\title{convert a sql database to an mgf file}
\usage{
sql_to_mgf(
  sqlite_path,
  expid,
  msLevel = NULL,
  spectrum_type = NULL,
  output_file
)
}
\arguments{
\item{sqlite_path}{'character()' the file path to a sql database.}

\item{expid}{'numeric()' experiment number of the data to convert to
an mgf file.}

\item{msLevel}{'numeric()' msLevel number of the data to convert to
an mgf file.}

\item{spectrum_type}{'character()' For flax ftms data it could be
"not_assembled" or "assembled".
For Qtof flax data it could be "intersect_single_energy"
or "merged_MSMS_all_energies".}

\item{output_file}{'character()' the output directory of the final mgf file.}
}
\value{
an mgf file with the data from the sqlite database.
}
\description{
sql_to_mgf() function first convert a sqlite database to a CompDb
object and then to a spectra object after that we extract the spectra
variables needed to use with MsBackendMgf package and finally we get that
mgf file of a given experiment with a spectrum_type in a sqlite database.
}
\author{
Ahlam Mentag
}
