% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_to_spectra.R
\name{sql_to_spectra}
\alias{sql_to_spectra}
\title{convert an sql database to a spectra object}
\usage{
sql_to_spectra(sql_path)
}
\arguments{
\item{sql_path}{the path to the sql database}
}
\value{
a spectra object of the sql database
}
\description{
convert an sql database to a spectra object
}
\author{
Ahlam Mentag
}
